--- prinz/src/main/scala/pl/touk/nussknacker/prinz/model/proxy/ProxiedInputModel.scala
@@ -1,56 +1,0 @@
-package pl.touk.nussknacker.prinz.model.proxy
-
-import pl.touk.nussknacker.prinz.model.{Model, ModelInstance, ModelMetadata, ModelName, ModelVersion}
-import pl.touk.nussknacker.engine.util.SynchronousExecutionContext.ctx
-import pl.touk.nussknacker.prinz.util.collection.immutable.VectorMultimap
-
-import scala.concurrent.Future
-
-class ProxiedInputModel(model: Model,
-                        private val proxiedParams: Iterable[ProxiedModelInputParam],
-                        private val compositeProxiedParams: Iterable[ProxiedModelCompositeInputParam[_ <: AnyRef]]) extends Model {
-
-  private val originalModelInstance = model.toModelInstance
-
-  private val proxiedName = new ModelName(s"Proxied-${model.getName.internal}")
-
-  private val modelMetadata = ModelMetadata(model.getName, model.getVersion, originalModelInstance.getSignature)
-
-  private val proxiedModelInstance = new ModelInstance(
-    originalModelInstance.model,
-    originalModelInstance.signatureProvider
-  ) {
-    override def run(inputMap: VectorMultimap[String, AnyRef]): ModelRunResult = {
-      val addInputParams = supplyNonProvidedInputs(inputMap)
-      val addComposedParams = addInputParams.flatMap(supplyNonProvidedComposedInputs)
-      addComposedParams.flatMap(originalModelInstance.run)
-    }
-  }
-
-  override def getName: ModelName = proxiedName
-
-  override def getVersion: ModelVersion = model.getVersion
-
-  override def toModelInstance: ModelInstance = proxiedModelInstance
-
-  private def supplyNonProvidedInputs(inputMap: VectorMultimap[String, AnyRef]): Future[VectorMultimap[String, AnyRef]] =
-    Future.sequence(
-      proxiedParams
-        .filter(inputsToBeReplacedIn(inputMap))
-        .map(_.supplyParamValue(modelMetadata))
-    ).map(addExtraInputsTo(inputMap))
-
-  private def supplyNonProvidedComposedInputs(inputMap: VectorMultimap[String, AnyRef]): Future[VectorMultimap[String, AnyRef]] =
-    Future.sequence(compositeProxiedParams
-      .map(_.supplyCompositeParamValues(modelMetadata))
-    )
-      .map(_.foldLeft(inputMap) { (acc, composedValues) =>
-        addExtraInputsTo(acc)(composedValues)
-      })
-
-  private def inputsToBeReplacedIn(inputMap: VectorMultimap[String, AnyRef]): ProxiedModelInputParam => Boolean =
-    param => !inputMap.containsKey(param.paramName.name)
-
-  private def addExtraInputsTo(inputMap: VectorMultimap[String, AnyRef])(extraInputs: Iterable[(String, AnyRef)]): VectorMultimap[String, AnyRef] =
-    extraInputs.foldLeft(inputMap) { (acc, value) => acc.add(value._1, value._2) }
-}
