+++ prinz/src/main/scala/pl/touk/nussknacker/prinz/model/proxy/api/ProxiedModelInputParam.scala
@@ -1,8 +1,9 @@
-package pl.touk.nussknacker.prinz.model.proxy
+package pl.touk.nussknacker.prinz.model.proxy.api
 
-import pl.touk.nussknacker.prinz.model.{ModelMetadata, SignatureName}
-import pl.touk.nussknacker.prinz.model.proxy.ProxiedModelInputParam.ParamSupplier
 import pl.touk.nussknacker.engine.util.SynchronousExecutionContext.ctx
+import pl.touk.nussknacker.prinz.model.proxy.api
+import pl.touk.nussknacker.prinz.model.proxy.api.ProxiedModelInputParam.ParamSupplier
+import pl.touk.nussknacker.prinz.model.{ModelMetadata, SignatureName}
 
 import scala.concurrent.Future
 
@@ -10,7 +11,7 @@ final case class ProxiedModelInputParam(paramName: SignatureName,
                                         paramSupplier: ParamSupplier) {
 
   def supplyParamValue(modelMetadata: ModelMetadata): Future[(String, AnyRef)] = {
-    val metadata = ModelInputParamMetadata(paramName, modelMetadata)
+    val metadata = api.ModelInputParamMetadata(paramName, modelMetadata)
     val resultFuture = paramSupplier(metadata)
     resultFuture.map { result =>
       (paramName.name, result)
